# skills/gcms-content-assistant/references/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GCMS Automation API",
    "version": "1.0.0",
    "description": "Automation API for enabled languages, media uploads, post categories, link categories, post and link draft previews, posts, pages, and links. Replace the server URL with GCMS_API_BASE for the target site."
  },
  "servers": [
    {
      "url": "https://example.com/api/admin/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
...
    "/page-bindings/preview": {
      "post": {
        "operationId": "pages.bindings.preview",
        "summary": "Preview a manifest or component binding against live data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object" }
            }
          }
        },
        "responses": {
          "200": { "description": "Resolved preview data without writing a revision" }
        }
      }
    },
...
    "/page-projects/{project_id}/capabilities/request": {
      "post": {
        "operationId": "pages.app.capabilities.request",
        "summary": "Request a manifest-declared app capability",
        "parameters": [
          { "$ref": "#/components/parameters/PageProjectId" },
          { "$ref": "#/components/parameters/PageIfMatch" },
          { "$ref": "#/components/parameters/PageIdempotencyKey" }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/PageCapabilityRequest"
        },
        "responses": {
          "201": { "description": "Capability request recorded" },
          "422": { "description": "Capability is undeclared or unavailable" }
        }
      }
    },
...
    "/page-projects/{project_id}/preview-url": {
      "post": {
        "operationId": "pages.preview",
        "summary": "Create a short-lived private preview for an exact revision/build",
        "parameters": [
          { "$ref": "#/components/parameters/PageProjectId" },
          { "$ref": "#/components/parameters/PageIfMatch" }
        ],
        "responses": { "201": { "description": "Signed no-store/noindex preview URL" } }
      }
    },
...
    "/page-projects/{project_id}/publish": {
      "post": {
        "operationId": "pages.publish",
        "summary": "Publish after Pilot-native target-bound confirmation",
        "description": "Without native confirmation returns unlock_required and an opaque page_challenge. The backend password and internal approval token never enter the model context. The native token is bound to operation/site/page/project/revision/build/live-data snapshot/ETag/request-id/subject.",
        "parameters": [
          { "$ref": "#/components/parameters/PageProjectId" },
          { "$ref": "#/components/parameters/PageIfMatch" },
          { "$ref": "#/components/parameters/PageIdempotencyKey" },
          { "$ref": "#/components/parameters/PageControlUnlock" }
        ],
        "responses": {
          "201": { "description": "Publication created" },
          "409": { "description": "Target/build conflict, build_stale, or native confirmation required" }
        }
      }
    },
...
    "/page-projects/{project_id}/rollback": {
      "post": {
        "operationId": "pages.rollback",
        "summary": "Rollback after Pilot-native target-bound confirmation",
        "description": "The same native boundary as publication binds operation/site/page/project/revision/build/live-data snapshot/ETag/request-id/subject; the AI never receives the password or internal approval token.",
        "parameters": [
          { "$ref": "#/components/parameters/PageProjectId" },
          { "$ref": "#/components/parameters/PageIfMatch" },
          { "$ref": "#/components/parameters/PageIdempotencyKey" },
          { "$ref": "#/components/parameters/PageControlUnlock" }
        ],
        "responses": {
          "201": { "description": "Rollback publication created" },
          "409": { "description": "Target/build conflict, build_stale, or native confirmation required" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "GCMS Access Token"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-GCMS-API-Key"
      }
    },
...
    "parameters": {
      "PageProjectId": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "minimum": 1 }
      },
...
      "PageIfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "Exact ETag from the latest project response. Never guess or overwrite after a conflict.",
        "schema": { "type": "string" }
      },
      "PageIdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Stable only for byte-equivalent retries of one logical request.",
        "schema": { "type": "string", "minLength": 8, "maxLength": 200 }
      },
      "PageControlUnlock": {
        "name": "X-GCMS-Control-Unlock",
        "in": "header",
        "required": false,
        "description": "Short-lived target-bound credential injected only by Pilot native UI. The AI must never receive, print, or persist it.",
        "schema": { "type": "string", "writeOnly": true }
      },
...
    },
...
    "requestBodies": {
      "PageCapabilityRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["capability"],
              "properties": {
                "capability": { "type": "string" },
                "config": { "type": "object", "additionalProperties": true }
              },
              "additionalProperties": false
            }
          }
        }
      },
...
    },
...
  }
}